docs: update Firestore default DB name in Terraform docs
The current docs incorrectly state that if TF_VAR_firestore_state_database
is not set, the default Firestore database (`(default)`) will be used.
In reality, the autoscaler-firestore module defaults to creating a dedicated
database named "memorystore-autoscaler-state". This commit updates the docs
to reflect that behavior, ensuring users know that by default, the Autoscaler
state will be stored in "memorystore-autoscaler-state".

Signed-off-by: Hitesh Narendra Dhruna <hitesh@dhruna.com>